+++ x/vpool/keeper/keeper.go
@@ -3,6 +3,9 @@ package keeper
 import (
 	"fmt"
 
+	"github.com/NibiruChain/nibiru/collections"
+	"github.com/NibiruChain/nibiru/collections/keys"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/tendermint/tendermint/libs/log"
@@ -17,9 +20,11 @@ func NewKeeper(
 	pricefeedKeeper types.PricefeedKeeper,
 ) Keeper {
 	return Keeper{
-		codec:           codec,
-		storeKey:        storeKey,
-		pricefeedKeeper: pricefeedKeeper,
+		codec:            codec,
+		storeKey:         storeKey,
+		pricefeedKeeper:  pricefeedKeeper,
+		Pools:            collections.NewMap[common.AssetPair, types.VPool](codec, storeKey, 0),
+		ReserveSnapshots: collections.NewMap[keys.Pair[common.AssetPair, keys.Uint64Key], types.ReserveSnapshot](codec, storeKey, 1),
 	}
 }
 
@@ -27,6 +32,9 @@ type Keeper struct {
 	codec           codec.BinaryCodec
 	storeKey        sdk.StoreKey
 	pricefeedKeeper types.PricefeedKeeper
+
+	Pools            collections.Map[common.AssetPair, types.VPool, *types.VPool]
+	ReserveSnapshots collections.Map[keys.Pair[common.AssetPair, keys.Uint64Key], types.ReserveSnapshot, *types.ReserveSnapshot]
 }
 
 func (k Keeper) Logger(ctx sdk.Context) log.Logger {
@@ -67,7 +75,7 @@ func (k Keeper) SwapBaseForQuote(
 		return sdk.Dec{}, types.ErrNoValidPrice.Wrapf("%s", pair.String())
 	}
 
-	pool, err := k.getPool(ctx, pair)
+	pool, err := k.Pools.Get(ctx, pair)
 	if err != nil {
 		return sdk.Dec{}, types.ErrPairNotSupported
 	}
@@ -158,7 +166,7 @@ func (k Keeper) SwapQuoteForBase(
 		return sdk.Dec{}, types.ErrNoValidPrice.Wrapf("%s", pair.String())
 	}
 
-	pool, err := k.getPool(ctx, pair)
+	pool, err := k.Pools.Get(ctx, pair)
 	if err != nil {
 		return sdk.Dec{}, types.ErrPairNotSupported
 	}
@@ -258,11 +266,12 @@ func (k Keeper) checkFluctuationLimitRatio(ctx sdk.Context, pool types.VPool) er
 		return nil
 	}
 
-	latestSnapshot, err := k.GetLatestReserveSnapshot(ctx, pool.Pair)
-	if err != nil {
+	it := k.ReserveSnapshots.Iterate(ctx, keys.NewRange[keys.Pair[common.AssetPair, keys.Uint64Key]]().Descending())
+	defer it.Close()
+	if !it.Valid() {
 		return fmt.Errorf("error getting last snapshot number for pair %s", pool.Pair)
 	}
-
+	latestSnapshot := it.Value()
 	if pool.IsOverFluctuationLimitInRelationWithSnapshot(latestSnapshot) {
 		return types.ErrOverFluctuationLimit
 	}
@@ -279,10 +288,10 @@ args:
   - pair: the asset pair
 
 ret:
-  - bool: whether or not the price has deviated from the oracle price beyond a spread ratio
+  - bool: whether the price has deviated from the oracle price beyond a spread ratio
 */
 func (k Keeper) IsOverSpreadLimit(ctx sdk.Context, pair common.AssetPair) bool {
-	pool, err := k.getPool(ctx, pair)
+	pool, err := k.Pools.Get(ctx, pair)
 	if err != nil {
 		panic(err)
 	}
@@ -306,7 +315,7 @@ ret:
   - sdk.Dec: The maintenance margin ratio for the pool
 */
 func (k Keeper) GetMaintenanceMarginRatio(ctx sdk.Context, pair common.AssetPair) sdk.Dec {
-	pool, err := k.getPool(ctx, pair)
+	pool, err := k.Pools.Get(ctx, pair)
 	if err != nil {
 		panic(err)
 	}
@@ -325,7 +334,7 @@ ret:
   - sdk.Dec: The maintenance margin ratio for the pool
 */
 func (k Keeper) GetMaxLeverage(ctx sdk.Context, pair common.AssetPair) sdk.Dec {
-	pool, err := k.getPool(ctx, pair)
+	pool, err := k.Pools.Get(ctx, pair)
 	if err != nil {
 		panic(err)
 	}