+++ x/vpool/keeper/prices_test.go
@@ -4,6 +4,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/NibiruChain/nibiru/collections/keys"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
@@ -247,21 +249,18 @@ func TestCalcTwap(t *testing.T) {
 					sdk.NewDec(10),
 					sdk.NewDec(90),
 					time.UnixMilli(10),
-					1,
 				),
 				types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					sdk.NewDec(10),
 					sdk.NewDec(85),
 					time.UnixMilli(20),
-					2,
 				),
 				types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					sdk.NewDec(10),
 					sdk.NewDec(95),
 					time.UnixMilli(30),
-					3,
 				),
 			},
 			currentBlockTime:   time.UnixMilli(30),
@@ -279,21 +278,18 @@ func TestCalcTwap(t *testing.T) {
 					sdk.NewDec(10),
 					sdk.NewDec(90),
 					time.UnixMilli(10),
-					1,
 				),
 				types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					sdk.NewDec(10),
 					sdk.NewDec(85),
 					time.UnixMilli(20),
-					2,
 				),
 				types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					sdk.NewDec(10),
 					sdk.NewDec(95),
 					time.UnixMilli(30),
-					3,
 				),
 			},
 			currentBlockTime:   time.UnixMilli(35),
@@ -303,21 +299,14 @@ func TestCalcTwap(t *testing.T) {
 			expectedPrice:      sdk.MustNewDecFromStr("8.895833333333333333"),
 		},
 		{
-			name: "spot price twap calc, t=[10,10]",
-			pair: common.Pair_BTC_NUSD,
-			reserveSnapshots: []types.ReserveSnapshot{
-				{
-					QuoteAssetReserve: sdk.NewDec(90),
-					BaseAssetReserve:  sdk.NewDec(10),
-					TimestampMs:       10,
-					BlockNumber:       1,
-				},
-			},
-			currentBlockTime:   time.UnixMilli(10),
+			name:               "spot price twap calc, t=[0,0]",
+			pair:               common.Pair_BTC_NUSD,
+			reserveSnapshots:   []types.ReserveSnapshot{},
+			currentBlockTime:   time.UnixMilli(0),
 			currentBlockHeight: 1,
-			lookbackInterval:   5 * time.Millisecond,
+			lookbackInterval:   0 * time.Millisecond,
 			twapCalcOption:     types.TwapCalcOption_SPOT,
-			expectedPrice:      sdk.NewDec(9),
+			expectedPrice:      sdk.NewDec(100),
 		},
 		{
 			name: "quote asset swap twap calc, add to pool, t=[10,30]",
@@ -328,14 +317,12 @@ func TestCalcTwap(t *testing.T) {
 					sdk.NewDec(10),
 					sdk.NewDec(30),
 					time.UnixMilli(10),
-					1,
 				),
 				types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					sdk.MustNewDecFromStr("7.5"),
 					sdk.NewDec(40),
 					time.UnixMilli(20),
-					2,
 				),
 			},
 			currentBlockTime:   time.UnixMilli(30),
@@ -355,14 +342,12 @@ func TestCalcTwap(t *testing.T) {
 					sdk.NewDec(10),
 					sdk.NewDec(60),
 					time.UnixMilli(10),
-					1,
 				),
 				types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					sdk.NewDec(12),
 					sdk.NewDec(50),
 					time.UnixMilli(20),
-					2,
 				),
 			},
 			currentBlockTime:   time.UnixMilli(30),
@@ -382,14 +367,12 @@ func TestCalcTwap(t *testing.T) {
 					sdk.NewDec(10),
 					sdk.NewDec(60),
 					time.UnixMilli(10),
-					1,
 				),
 				types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					sdk.NewDec(20),
 					sdk.NewDec(30),
 					time.UnixMilli(20),
-					2,
 				),
 			},
 			currentBlockTime:   time.UnixMilli(30),
@@ -409,14 +392,12 @@ func TestCalcTwap(t *testing.T) {
 					sdk.NewDec(10),
 					sdk.NewDec(60),
 					time.UnixMilli(10),
-					1,
 				),
 				types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					sdk.NewDec(8),
 					sdk.NewDec(75),
 					time.UnixMilli(20),
-					2,
 				),
 			},
 			currentBlockTime:   time.UnixMilli(30),
@@ -434,7 +415,7 @@ func TestCalcTwap(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			vpoolKeeper, ctx := VpoolKeeper(t,
 				mock.NewMockPricefeedKeeper(gomock.NewController(t)))
-			ctx = ctx.WithBlockTime(time.UnixMilli(0)).WithBlockHeight(1)
+			ctx = ctx.WithBlockTime(time.UnixMilli(0))
 
 			t.Log("Create an empty pool for the first block")
 			vpoolKeeper.CreatePool(
@@ -463,18 +444,14 @@ func TestCalcTwap(t *testing.T) {
 			)
 
 			for _, snapshot := range tc.reserveSnapshots {
-				ctx = ctx.WithBlockHeight(snapshot.BlockNumber).WithBlockTime(time.UnixMilli(snapshot.TimestampMs))
+				ctx = ctx.WithBlockTime(time.UnixMilli(snapshot.TimestampMs))
 				snapshot := types.NewReserveSnapshot(
 					common.Pair_BTC_NUSD,
 					snapshot.BaseAssetReserve,
 					snapshot.QuoteAssetReserve,
 					ctx.BlockTime(),
-					ctx.BlockHeight(),
 				)
-				vpoolKeeper.SaveSnapshot(
-					ctx,
-					snapshot,
-				)
+				vpoolKeeper.ReserveSnapshots.Insert(ctx, keys.Join(snapshot.Pair, keys.Uint64(uint64(snapshot.TimestampMs))), snapshot)
 			}
 			ctx = ctx.WithBlockTime(tc.currentBlockTime).WithBlockHeight(tc.currentBlockHeight)
 			price, err := vpoolKeeper.calcTwap(ctx,